requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.4
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.4
django-allauth==0.57.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-heroku==0.3.1
django-timezone-field==6.0.1
django-widget-tweaks==1.5.0
gunicorn==21.2.0
idna==3.4
kombu==5.3.3
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
prompt-toolkit==3.0.41
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.10
whitenoise==6.6.0
